requirements.txt
# vLLM Benchmark Requirements
# ============================

# Core vLLM Framework
vllm>=0.8.0,<0.9.0              # Main vLLM library for LLM inference
torch>=2.6.0,<3.0.0             # PyTorch for deep learning
transformers>=4.51.0,<5.0.0     # Hugging Face transformers for tokenizers and models (Qwen3 support)

# AI/ML Libraries
numpy>=1.26.4,<2.0.0            # Numerical computing
matplotlib>=3.9.0,<4.0.0        # Plotting and visualization
pandas>=2.2.0,<3.0.0            # Data manipulation and analysis

# HTTP Client and Networking
openai>=1.55.0,<2.0.0           # OpenAI API client for HTTP requests
requests>=2.32.0,<3.0.0         # HTTP library
httpx>=0.28.0,<1.0.0            # Modern async HTTP client
aiohttp>=3.11.0,<4.0.0          # Async HTTP client/server framework

# Async Programming
anyio>=4.6.0,<5.0.0            # Async networking and concurrency library

# Data Formats and Serialization
PyYAML>=6.0.0,<7.0.0           # YAML parser and emitter
pydantic>=2.10.0,<3.0.0        # Data validation using Python type hints

# System Monitoring and GPU Management
psutil>=6.1.0,<7.0.0           # System and process monitoring
pynvml>=12.0.0,<13.0.0         # NVIDIA GPU monitoring
nvidia-ml-py>=12.575.0,<13.0.0 # NVIDIA Machine Learning library

# SSH and Tunneling
paramiko>=3.5.0,<4.0.0         # SSH client
sshtunnel>=0.4.0,<1.0.0        # SSH tunnel support

# Security and Cryptography
cryptography>=44.0.0,<45.0.0   # Cryptographic recipes and primitives
bcrypt>=4.2.0,<5.0.0           # Password hashing

# File Processing and Storage
huggingface-hub>=0.28.0,<1.0.0 # Hugging Face model hub access
datasets>=3.3.0,<4.0.0         # Hugging Face datasets library
safetensors>=0.5.0,<1.0.0      # Safe tensor serialization

# Utilities and Data Processing
tqdm>=4.67.0,<5.0.0            # Progress bars
regex>=2024.11.0               # Regular expressions
tokenizers>=0.21.0,<1.0.0     # Fast tokenizers
filelock>=3.17.0,<4.0.0       # Platform independent file locking

# Development and Debugging
setuptools>=75.8.0             # Package development tools

# Data Types and Validation
typing_extensions>=4.12.0,<5.0.0  # Backported typing features

# Date and Time
python-dateutil>=2.9.0,<3.0.0  # Date utilities
pytz>=2025.1                    # Timezone library

# Core Python Libraries (usually included)
certifi>=2024.8.30             # Certificate bundle
urllib3>=2.2.0,<3.0.0          # HTTP library
idna>=3.10                      # Internationalized domain names
charset-normalizer>=3.4.0,<4.0.0  # Character encoding detection

# Additional Dependencies (auto-installed by main packages)
aiohappyeyeballs>=2.4.0,<3.0.0 # IPv6 happy eyeballs for aiohttp
aiosignal>=1.3.0,<2.0.0        # Async signal handling
annotated-types>=0.7.0,<1.0.0  # Type annotations
attrs>=24.2.0,<25.0.0          # Classes without boilerplate
cffi>=1.17.0,<2.0.0           # C Foreign Function Interface
contourpy>=1.3.0,<2.0.0       # Contouring library for matplotlib
cycler>=0.12.0,<1.0.0         # Composable style cycles
dill>=0.3.8,<1.0.0            # Serialize almost anything
distro>=1.9.0,<2.0.0          # Linux distribution detection
fonttools>=4.55.0,<5.0.0      # Font tools for matplotlib
frozenlist>=1.5.0,<2.0.0      # Immutable list
fsspec>=2024.12.0             # File system specification
h11>=0.14.0,<1.0.0            # HTTP/1.1 protocol
httpcore>=1.0.0,<2.0.0        # HTTP core
jiter>=0.8.0,<1.0.0           # JSON iterator
Jinja2>=3.1.0,<4.0.0          # Template engine
kiwisolver>=1.4.0,<2.0.0      # Fast constraint solver
MarkupSafe>=3.0.0,<4.0.0      # XML/HTML/XHTML markup safe string
mpmath>=1.3.0,<2.0.0          # Multiple precision arithmetic
multidict>=6.1.0,<7.0.0       # Multi-dict implementation
multiprocess>=0.70.0,<1.0.0   # Better multiprocessing
networkx>=3.4.0,<4.0.0        # Network analysis
packaging>=24.2,<25.0.0       # Core packaging utilities
pillow>=11.0.0,<12.0.0        # Python Imaging Library
propcache>=0.2.0,<1.0.0       # Property caching
pyarrow>=19.0.0,<20.0.0       # Apache Arrow for data
pycparser>=2.22,<3.0.0        # C parser
pydantic_core>=2.27.0,<3.0.0  # Pydantic core
PyNaCl>=1.5.0,<2.0.0         # Cryptography library
pyparsing>=3.2.0,<4.0.0      # Parsing library
six>=1.17.0,<2.0.0           # Python 2 and 3 compatibility utilities
sniffio>=1.3.0,<2.0.0        # Async library detection
sympy>=1.13.0,<2.0.0         # Symbolic mathematics
tzdata>=2025.1                # Timezone database
xxhash>=3.5.0,<4.0.0         # Fast hash algorithm
yarl>=1.18.0,<2.0.0          # Yet another URL library